+++ CHANGELOG.md
@@ -22,7 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### New
 
 - The sleepit helper now performs signal handling, to enable better tests.
-- Add CI: build and test with GitHub Actions, for platforms: Linux, macOS.
+- Add CI: build and test with GitHub Actions, for platforms: Linux, macOS, Windows.
+- Add basic and experimental support for Windows. Untested: signal handling.
 
 ## [v0.6.0] - [2021-01-17]